+++ waffle-chai/src/matchers/changeEtherBalances.ts
@@ -1,4 +1,5 @@
 import {BigNumber, BigNumberish, providers} from 'ethers';
+import {callPromise} from '../call-promise';
 import {getAddressOf, Account} from './misc/account';
 import {BalanceChangeOptions, getAddresses, getBalances} from './misc/balance';
 
@@ -9,47 +10,43 @@ export function supportChangeEtherBalances(Assertion: Chai.AssertionStatic) {
     balanceChanges: BigNumberish[],
     options: BalanceChangeOptions
   ) {
-    const subject = this._obj;
-
-    const derivedPromise = Promise.all([
-      getBalanceChanges(subject, accounts, options),
-      getAddresses(accounts)
-    ]).then(
-      ([actualChanges, accountAddresses]) => {
-        this.assert(
-          actualChanges.every((change, ind) =>
-            change.eq(BigNumber.from(balanceChanges[ind]))
-          ),
-          `Expected ${accountAddresses} to change balance by ${balanceChanges} wei, ` +
-            `but it has changed by ${actualChanges} wei`,
-          `Expected ${accountAddresses} to not change balance by ${balanceChanges} wei,`,
-          balanceChanges.map((balanceChange) => balanceChange.toString()),
-          actualChanges.map((actualChange) => actualChange.toString())
-        );
+    callPromise(this);
+    const isNegated = this.__flags.negate === true;
+    const derivedPromise = this.callPromise.then(() => {
+      if (!('txResponse' in this)) {
+        throw new Error('The changeEtherBalances matcher must be called on a transaction');
       }
-    );
+      return Promise.all([
+        getBalanceChanges(this.txResponse, accounts, options),
+        getAddresses(accounts)
+      ]);
+    }).then(([actualChanges, accountAddresses]: [BigNumber[], string[]]) => {
+      const isCurrentlyNegated = this.__flags.negate === true;
+      this.__flags.negate = isNegated;
+      this.assert(
+        actualChanges.every((change, ind) =>
+          change.eq(BigNumber.from(balanceChanges[ind]))
+        ),
+        `Expected ${accountAddresses} to change balance by ${balanceChanges} wei, ` +
+          `but it has changed by ${actualChanges} wei`,
+        `Expected ${accountAddresses} to not change balance by ${balanceChanges} wei,`,
+        balanceChanges.map((balanceChange) => balanceChange.toString()),
+        actualChanges.map((actualChange) => actualChange.toString())
+      );
+      this.__flags.negate = isCurrentlyNegated;
+    });
     this.then = derivedPromise.then.bind(derivedPromise);
     this.catch = derivedPromise.catch.bind(derivedPromise);
-    this.promise = derivedPromise;
+    this.callPromise = derivedPromise;
     return this;
   });
 }
 
 export async function getBalanceChanges(
-  transaction:
-  | providers.TransactionResponse
-  | (() => Promise<providers.TransactionResponse> | providers.TransactionResponse),
+  txResponse: providers.TransactionResponse,
   accounts: Account[],
   options: BalanceChangeOptions
 ) {
-  let txResponse: providers.TransactionResponse;
-
-  if (typeof transaction === 'function') {
-    txResponse = await transaction();
-  } else {
-    txResponse = transaction;
-  }
-
   const txReceipt = await txResponse.wait();
   const txBlockNumber = txReceipt.blockNumber;
 
@@ -73,6 +70,11 @@ async function getTxFees(
         const gasPrice = txResponse.gasPrice ?? txReceipt.effectiveGasPrice;
         const gasUsed = txReceipt.gasUsed;
         const txFee = gasPrice.mul(gasUsed);
+        const provider = account.provider as any;
+        if (typeof provider.getL1Fee === 'function') {
+          const l1Fee = await provider.getL1Fee(txReceipt.transactionHash);
+          return txFee.add(l1Fee);
+        }
 
         return txFee;
       }