+++ waffle-chai/src/matchers/reverted.ts
@@ -1,16 +1,8 @@
+import {callPromise} from '../call-promise';
+
 export function supportReverted(Assertion: Chai.AssertionStatic) {
   Assertion.addProperty('reverted', function (this: any) {
-    const promise = this._obj;
-    const onSuccess = (value: any) => {
-      this.assert(
-        false,
-        'Expected transaction to be reverted',
-        'Expected transaction NOT to be reverted',
-        'Transaction reverted.',
-        'Transaction NOT reverted.'
-      );
-      return value;
-    };
+    callPromise(this);
     const onError = (error: any) => {
       const message = (error instanceof Object && 'message' in error) ? error.message : JSON.stringify(error);
       const isReverted = message.search('revert') >= 0;
@@ -19,15 +11,24 @@ export function supportReverted(Assertion: Chai.AssertionStatic) {
       this.assert(
         isReverted || isThrown || isError,
         `Expected transaction to be reverted, but other exception was thrown: ${error}`,
-        'Expected transaction NOT to be reverted',
+        `Expected transaction NOT to be reverted, but it was reverted with "${message}"`,
         'Transaction reverted.',
         error
       );
       return error;
     };
-    const derivedPromise = promise.then(onSuccess, onError);
-    this.then = derivedPromise.then.bind(derivedPromise);
-    this.catch = derivedPromise.catch.bind(derivedPromise);
+
+    const assertNotReverted = () => this.assert(
+      false,
+      'Expected transaction to be reverted',
+      'Expected transaction NOT to be reverted',
+      'Transaction reverted.',
+      'Transaction NOT reverted.'
+    );
+
+    this.callPromise = this.callPromise.then(assertNotReverted, onError);
+    this.then = this.callPromise.then.bind(this.callPromise);
+    this.catch = this.callPromise.catch.bind(this.callPromise);
     return this;
   });
 }