+++ waffle-hardhat/test/constants.ts
@@ -1,0 +1,2 @@
+export const TX_GAS = 21000; // Gas used by a single, non-contract transaction.
+export const BASE_FEE_PER_GAS = 875000000;