+++ waffle-optimism/test/events.test.ts
@@ -1,0 +1,8 @@
+import {eventsTest} from '@ethereum-waffle/chai/test';
+import {getOptimismProvider} from '../src/provider';
+
+describe('Optimism: changeEtherBalance matcher', () => {
+  const provider = getOptimismProvider();
+
+  eventsTest(provider);
+});